agents/requirements.txt
# Core
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
pydantic-settings==2.5.2
httpx==0.27.2

# LangGraph / LangChain
langgraph==0.2.53
langchain==0.3.7
langchain-openai==0.2.9
langchain-core==0.3.19

# Kafka
confluent-kafka==2.6.1

# Redis
redis==5.2.1

# Observability
opentelemetry-api==1.28.2
opentelemetry-sdk==1.28.2
opentelemetry-exporter-otlp==1.28.2
opentelemetry-instrumentation-fastapi==0.49b2
opentelemetry-instrumentation-httpx==0.49b2
opentelemetry-instrumentation-redis==0.49b2

# Utilities
python-dotenv==1.0.1
structlog==24.4.0
tenacity==9.0.0
